% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_nodes}
\alias{decorate_nodes}
\title{Graph Decorator Functions}
\usage{
decorate_nodes(g, data = NULL, func = NULL, ...)
}
\description{
Decorate graphs with anonymous functions to be passed to the graph decorator operator `%@%`.
}
\details{
For examples and tutorials, please see the repository: \url{https://github.com/esteinig/netviewr}
Example function
}
