% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R
\name{find_contributors}
\alias{find_contributors}
\title{Identification of Key Contributors}
\usage{
find_contributors(matrix, metaData, paran_iterations=100, paran_centile=99, dist=FALSE, verbose=TRUE)
}
\arguments{
\item{matrix}{Pairwise relationship / correlation / similarity matrix between samples (N x N)}

\item{paran_iterations}{Iterations for computing the number of significant components using Horn's Parallel Analysis in Paran [ int, 100 ]}

\item{paran_centile}{Significance level (centile used in estimating bias) in Paran [ int, 99 ]}

\item{dist}{Input matrix is a distance matrix; converted to similarity matrix by subtracting from 1 [ bool, FALSE ]}

\item{verbose}{Print computation status and results of Horn's Parallel Analysis [ bool, TRUE ]}
}
\value{
Tibble containing ordered genetic contribution scores (GCS)
}
\description{
Identification of key contributors using Horn's Parallel Analysis, Eigenvalue Decomposition and
calculation of the Genetic Contribution Score (GCS) as described in Neuditschko et al. (2017).
}
\details{
For examples and tutorials, please see the Repository: \url{https://github.com/esteinig/netviewr}
}
