% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%@\%}
\alias{\%@\%}
\title{Graph Decorator Operator
------------------------}
\usage{
`igraph::sample_gnm(n=10, m=15) %@% netviewr::node_size(size=5)`
}
\arguments{
\item{g}{graph object or list of graph objects to decorate}

\item{f}{decorator function to decorate the graph with, or a tibble}
}
\value{
Graph object or list of graph objects with decorated attributes
}
\description{
Decorate a graph object (g) a decorator function (f) or a tibble with meta data
for nodes (length: nodes)
Functions can be defined manually as functions that return an anonymous function.
}
