% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{meta_operator}
\alias{meta_operator}
\alias{\%m\%}
\title{Meta Operator (Column Concatenation Operator)}
\usage{
t1 \%m\% t2
}
\arguments{
\item{t1}{data frame or tibble containing some data}

\item{t2}{data frame or tibble containing some other data to merge}
}
\value{
Merged data frame by columns of t1 and t2
}
\description{
Adds a Tibble to the graph attribute `meta` (or concatenates to existing Tibble)
for use in subsequent graph decorators that might use the data argument to select
columns from Tibble:
}
