% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netview.R
\name{netview}
\alias{netview}
\title{Netview R}
\usage{
netview(dist, k=1:60, mutual=T, weights=T, mst=F, algorithm='cover_tree')
}
\arguments{
\item{dist}{Symmetrical distance matrix for NetView (N x N) [ matrix ]}

\item{k}{Range of parameter k for mutual k-nearest-neighbour search [ int vector, 10:60 ]}

\item{mutual}{Construct mutual nearest neighbor graph instead of nearest neighbor graph [ bool, TRUE ]}

\item{weights}{Weight of edges is mapped to pairwise distance in input matrix [ bool, TRUE ]}

\item{mst}{Include edges of the minimum spanning tree associated with the data [ bool, FALSE ]}

\item{algorithm}{Algorithm for mutual nearest neighbour search [ char cover_tree ]}
}
\value{
List of graph objects (igraph)
}
\description{
Generate mutual nearest-neighbour graphs for analysis of population structure and visualization with iGraph.
}
\details{
For examples and tutorials, please see the repository: \url{https://github.com/esteinig/netviewr}
}
