% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_kselect}
\alias{plot_kselect}
\title{K-selection Plot}
\usage{
plot_kselect(graphs)
}
\arguments{
\item{graphs}{vector/list of igraph objects at increasing k-nearest-neighbor values from netviewr::netview}
}
\description{
K-selection Plot
}
