% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_netview}
\alias{plot_netview}
\title{Netview Plot}
\usage{
plot_netview(
  g,
  legend = NULL,
  legend_position = "topleft",
  legend_size = 0.7,
  text = NULL,
  text_size = 1,
  text_color = "black",
  text_font = 1,
  ncol = NULL,
  nrow = NULL,
  dev_off = F,
  ...
)
}
\arguments{
\item{g}{igraph object: the igraph object to plot}

\item{legend}{char: column name from graph object attribute `node_data` to plot a legend (if data added with decorator)}

\item{legend_position}{char: legend position}

\item{legend_size}{numeric: legend size}

\item{text}{char: title text}

\item{text_size}{numeric: title size}

\item{text_color}{char: title color}

\item{text_font}{numeric: title font}

\item{ncol}{numeric: number of cols for panel (max 4 panels)}

\item{nrow}{numeric: number of rows for panel (max 4 panels)}

\item{dev_off}{bool: device off}

\item{...}{any other parameter passed to igraph::pot.igraph}
}
\description{
Netview Plot
}
