% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plt_netview}
\alias{plt_netview}
\title{Helper function}
\usage{
plt_netview(
  g,
  legend = NULL,
  legend_position = "topleft",
  legend_size = 0.7,
  text = NULL,
  text_size = 1,
  text_color = "black",
  text_font = 1,
  ...
)
}
\description{
Helper function
}
