% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_dist}
\alias{read_dist}
\title{Matrix Reader}
\usage{
read_dist(file = "", diag = FALSE, names = paste("X", 1:n, sep = ""), ...)
}
\arguments{
\item{file}{char: file path to matrix - no header, no rownames}

\item{diag}{bool: whether dist matrix is diagonal only}

\item{names}{vec: vector of row and column names}

\item{...}{any other parameter passed to `scan` function to read file}
}
\description{
Matrix Reader
}
